panel-ditem-editor: avoid sign-compare warning

+++ gnome-panel/panel-ditem-editor.c
@@ -1437,7 +1437,7 @@ response_cb (GtkDialog *dialog,
 static void
 panel_ditem_editor_revert (PanelDItemEditor *dialog)
 {
-	int       i;
+	gulong    i;
 	char     *string;
 	gboolean  boolean;
 	GKeyFile *key_file;
@@ -1513,7 +1513,7 @@ panel_ditem_editor_revert (PanelDItemEditor *dialog)
 static void
 panel_ditem_editor_set_revert (PanelDItemEditor *dialog)
 {
-	int       i;
+	gulong    i;
 	char     *string;
 	gboolean  boolean;
 	GKeyFile *key_file;
